pyproject.toml
[project]
name = "world_model_lens"
version = "0.2.0"
description = "Backend-agnostic interpretability library for ANY world model"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "World Model Lens Team" },
]
keywords = [
    "world-models",
    "interpretability",
    "reinforcement-learning",
    "dreamer",
    "rl",
    "video-prediction",
    "planning",
    "robotics",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "numpy",
    "scipy",
    "scikit-learn",
    "pandas",
    "matplotlib",
    "plotly",
    "tqdm",
    "pyyaml",
    "rich",
    "typer",
    "einops",
    "huggingface_hub",
    "gymnasium",
]

[project.optional-dependencies]
viz = ["plotly", "kaleido"]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
    "black",
    "hatch",
    "flake8",
]
docs = [
    "sphinx",
    "sphinx-autodoc-typehints",
    "myst-parser",
    "pydata-sphinx-theme",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "slowapi>=0.1.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=5.9.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "prometheus-client>=0.17.0",
]

[project.scripts]
wml = "world_model_lens.cli.commands:app"

[project.urls]
Homepage = "https://github.com/Bhavith-Chandra/WorldModelLens"
Documentation = "https://worldmodellens.readthedocs.io"
Repository = "https://github.com/Bhavith-Chandra/WorldModelLens"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["world_model_lens"]

[tool.hatch.build.targets.sdist]
include = [
    "/world_model_lens",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short -m 'not network'"
pythonpath = ["."]
markers = [
    "network: marks tests that require a network connection and HuggingFace access (deselect with '-m \"not network\"')",
]
